+++ functions/public/Get-PSProjectGitStatus.ps1
@@ -4,6 +4,10 @@ Function Get-PSProjectGitStatus {
     [OutputType('PSProjectGit')]
     Param( )
 
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Using PowerShell Host $($Host.Name)"
+
     if (Test-Path .git) {
 
         $remotes = _getRemote | Where-Object { $_.mode -eq 'push' } |
@@ -21,4 +25,5 @@ Function Get-PSProjectGitStatus {
     else {
         Write-Verbose "You must run this command from the root of a git repository."
     }
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
 }