+++ functions/public/New-PSProjectStatus.ps1
@@ -29,45 +29,62 @@ Function New-PSProjectStatus {
         [version]$ProjectVersion,
 
         [Parameter(HelpMessage = "Enter an optional comment. This could be git tag, or an indication about the type of project.")]
-        [String]$Comment
+        [String]$Comment,
+
+        [Parameter(HelpMessage = "Overwrite an existing PSProjectStatus file.")]
+        [switch]$Force
     )
 
-    Write-Verbose "Starting $($MyInvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Using PowerShell Host $($Host.Name)"
 
     $exclude = "Verbose", "WhatIf", "Confirm", "ErrorAction", "Debug",
     "WarningAction", "WarningVariable", "ErrorVariable","InformationAction",
-    "InformationVariable"
+    "InformationVariable","ProgressAction","Force"
 
-    Write-Verbose "Creating a new instance of the PSProject class"
-    $new = [psproject]::New()
-    $new | Select-Object * | Out-String | Write-Debug
-    Write-Verbose "Using schema path $jsonSchema"
     #convert the path to a filesystem path to avoid using PSDrive references
-    $new.Path = Convert-Path $Path
-    Write-Verbose "Using path $Path"
+    $cPath = Convert-Path $Path
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Using path $Path"
+
+    #Don't overwrite existing file unless -Force is specified
+    $json = Join-Path -Path $cPath -ChildPath psproject.json
+
+    if ((Test-Path -path $json) -AND (-Not $Force)) {
+        Write-Warning "An existing project status file was found. You must either delete the file or re-run this command with -Force."
+        #bail out of the command
+        Return
+    }
+
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Creating a new instance of the PSProject class"
+    $new = [psproject]::New()
+    $new.Path = $cPath
+    $new | Select-Object * | Out-String | Write-Debug
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Using schema path $jsonSchema"
+
     #set the instance properties using parameter values from this command
     $PSBoundParameters.GetEnumerator() | Where-Object { $exclude -NotContains $_.key } |
     ForEach-Object {
-        Write-Verbose "Setting property $($_.key)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Setting property $($_.key)"
         $new.$($_.key) = $_.value
     }
 
     Write-Verbose "Testing for .git"
     If (Test-Path .git) {
         $branch = git branch --show-current
-        Write-Verbose "Detected git branch $branch"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Detected git branch $branch"
         $new.GitBranch = $branch
 
         #get git remote
         $new.RemoteRepository = _getRemote
     }
     else {
-        Write-Verbose "No git branch detected"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] No git branch detected"
     }
     if ($PSCmdlet.ShouldProcess($Name)) {
         $new
         $new.Save()
     }
-    Write-Verbose "Ending $($MyInvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
 
 }